+++ bpa/src/dispatcher/reassemble.rs
@@ -3,46 +3,41 @@ use tracing::debug;
 
 use super::Dispatcher;
 use crate::{
-    bundle::{Bundle, BundleMetadata, BundleStatus, ReadOnlyMetadata},
+    bundle::{Bundle, BundleMetadata, BundleStatus},
     storage::adu_reassembly::ReassemblyResult,
 };
 
 impl Dispatcher {
     pub async fn reassemble(&self, mut bundle: Bundle) {
-        let (storage_name, data, received_at) = match self.store.adu_reassemble(&bundle).await {
-            ReassemblyResult::NotReady => {
-                let status = BundleStatus::AduFragment {
-                    source: bundle.bundle.primary.id.source.clone(),
-                    timestamp: bundle.bundle.primary.id.timestamp.clone(),
-                };
-                self.store.update_status(&mut bundle, &status).await;
-                return self.store.watch_bundle(bundle).await;
-            }
-            ReassemblyResult::Failed => {
-                debug!(
-                    "Fragment reassembly failed for bundle {}",
-                    bundle.bundle.primary.id
-                );
-                return;
-            }
-            ReassemblyResult::Done {
-                storage_name,
-                data,
-                received_at,
-            } => (storage_name, data, received_at),
-        };
+        let (storage_name, data, received_at, origin) =
+            match self.store.adu_reassemble(&bundle).await {
+                ReassemblyResult::NotReady => {
+                    let status = BundleStatus::AduFragment {
+                        source: bundle.bundle.primary.id.source.clone(),
+                        timestamp: bundle.bundle.primary.id.timestamp.clone(),
+                    };
+                    self.store.update_status(&mut bundle, &status).await;
+                    return self.store.watch_bundle(bundle).await;
+                }
+                ReassemblyResult::Failed => {
+                    debug!(
+                        "Fragment reassembly failed for bundle {}",
+                        bundle.bundle.primary.id
+                    );
+                    return;
+                }
+                ReassemblyResult::Done {
+                    storage_name,
+                    data,
+                    received_at,
+                    origin,
+                } => (storage_name, data, received_at, origin),
+            };
 
         metrics::counter!("bpa.bundle.reassembled").increment(1);
 
-        let metadata = BundleMetadata {
-            storage_name: Some(storage_name.clone()),
-            status: BundleStatus::New,
-            read_only: ReadOnlyMetadata {
-                received_at,
-                ..Default::default()
-            },
-            ..Default::default()
-        };
+        let mut metadata = BundleMetadata::new(received_at, origin);
+        metadata.storage_name = Some(storage_name.clone());
 
         // TODO: Just push the entire bundle into the stream
         let (tx, rx) = hardy_async::channel::bounded(1);